+++ kotori/daq/services/__init__.py
@@ -2,7 +2,11 @@
 # (c) 2015-2016 Andreas Motl, Elmyra UG <andreas.motl@elmyra.de>
 from twisted.application.service import MultiService
 
-class RootService(MultiService):
+class MultiServiceMixin(object):
+    def registerService(self, service):
+        service.setServiceParent(self)
+
+class RootService(MultiService, MultiServiceMixin):
 
     def __init__(self, settings=None):
         MultiService.__init__(self)
@@ -12,6 +16,3 @@ class RootService(MultiService):
         service = service_class(*args, **kwargs)
         self.registerService(service)
         return service
-
-    def registerService(self, service):
-        service.setServiceParent(self)